wip: go from static mut to struct
This will allow more idomatic rust, allows the borrow checker to enforce
certain invariants, and removes large swaths of unsafe code. Now
technically these bits of unsafe aren't necessarily unsafe, but it still
helps clean up the code for *actual* unsafe review.

One current struggle is `Packet`s want to be able to add statistics
about how much memory has been freed, which before requires just
mutating a static global. However, if the `Engine` becomes a struct,
there is no way for a packet to have access to the engine. This will
take a little refactor to create some kind of allocator.

Also un-implemented as of yet is that `Engine` is not a singleton. This
is an easy fix, it's just not yet done.